Re-introduces custom image base layers for OpenLayers
This was possible in some old version of Maps already but got broken a long time ago. This fixes all issues and re-introduces the feature to map. Custom layers have to be defined in a Layer namespace. All layers will be saved in the database, therefore running update.php is required for this to work.

I will optimize some more and look at the database stuff, there could be some issues here. I originally wrote this for OpenLayers 1.1. but abandoned it for some months now, so it was not that simple merging it into the current master.

Change-Id: Ife7975abdaec1f683ea09c419dc91c8016b6bab1